Refactor `graph_traversal` file
