+++ bw_graph_tools/graph_traversal/new_node_each_visit.py
@@ -1,8 +1,6 @@
 import warnings
-from dataclasses import dataclass
-from functools import lru_cache
 from heapq import heappop, heappush
-from typing import Optional, Dict, List
+from typing import Optional, Dict
 
 import matrix_utils as mu
 import numpy as np
@@ -14,166 +12,9 @@ try:
 except ImportError:
     databases = {}
 
-from .matrix_tools import guess_production_exchanges
-
-
-class CachingSolver:
-    """Class which caches expensive linear algebra solution vectors."""
-
-    def __init__(self, lca: LCA):
-        self.lca = lca
-
-    @lru_cache(maxsize=8096)
-    def calculate(self, index: int) -> np.ndarray:
-        """Compute cumulative LCA score for one unit of a given activity"""
-        self.lca.demand_array[:] = 0
-        self.lca.demand_array[index] = 1
-        return self.lca.solve_linear_system()
-
-    def __call__(self, index: int, amount: float) -> np.ndarray:
-        return self.calculate(index) * amount
-
-
-class Counter:
-    """Custom counter to have easy access to current value"""
-
-    def __init__(self):
-        self.value = -1
-
-    def __next__(self):
-        self.value += 1
-        return self.value
-
-    def __gt__(self, other):
-        return self.value > other
-
-
-@dataclass
-class Node:
-    """
-    A visited activity in a supply chain graph. Although our graph is cyclic, we treat each
-    activity as a separate node every time we visit it.
-
-    Parameters
-    ----------
-    unique_id : int
-        A unique integer id for this visit to this activity node
-    activity_datapackage_id : int
-        The id that identifies this activity in the datapackage, and hence in the database
-    activity_index : int
-        The technosphere matrix column index of this activity
-    reference_product_datapackage_id : int
-        The id that identifies the reference product of this activity in the datapackage
-    reference_product_index : int
-        The technosphere matrix row index of this activity's reference product
-    reference_product_production_amount : float
-        The *net* production amount of this activity's reference product
-    depth : int
-        Depth in the supply chain graph, starting from 0 as the functional unit
-    supply_amount : float
-        The amount of the *activity* (not reference product!) needed to supply the demand from the
-        requesting supply chain edge.
-    cumulative_score : float
-        Total LCIA score attributed to `supply_amount` of this activity, including impacts from
-        direct emissions.
-    direct_emissions_score : float
-        Total LCIA score attributed only to the direct characterized biosphere flows of
-        `supply_amount` of this activity.
-    direct_emissions_score_outside_specific_flows : float
-        The score attributable to *direct emissions* of this node which isn't broken out into
-        separate `Flow` objects.
-    remaining_cumulative_score_outside_specific_flows : float
-        The *cumulative* score of this node, including direct emissions, which isn't broken out
-        into separate `Flow` objects.
-    terminal : bool
-        Boolean flag indicating whether graph traversal was cutoff at this node
-
-    """
-
-    unique_id: int
-    activity_datapackage_id: int
-    activity_index: int
-    reference_product_datapackage_id: int
-    reference_product_index: int
-    reference_product_production_amount: float
-    depth: int
-    supply_amount: float
-    cumulative_score: float
-    direct_emissions_score: float
-    direct_emissions_score_outside_specific_flows: float = 0.0
-    remaining_cumulative_score_outside_specific_flows: float = 0.0
-    terminal: bool = False
-
-    def __lt__(self, other):
-        # Needed for sorting
-        return self.cumulative_score < other.cumulative_score
-
-
-@dataclass
-class Edge:
-    """
-    An edge between two *activities*. The `amount` is the amount of the product demanded by the
-    `consumer`.
-
-    Parameters
-    ----------
-    consumer_index : int
-        The matrix column index of the consuming activity
-    consumer_unique_id : int
-        The traversal-specific unique id of the consuming activity
-    producer_index : int
-        The matrix column index of the producing activity
-    producer_unique_id : int
-        The traversal-specific unique id of the producing activity
-    product_index : int
-        The matrix row index of the consumed product
-    amount : float
-        The amount of the product demanded by the consumer. Not scaled to producer production
-        amount.
-    """
-
-    consumer_index: int
-    consumer_unique_id: int
-    producer_index: int
-    producer_unique_id: int
-    product_index: int
-    amount: float
-
-
-@dataclass
-class Flow:
-    """
-    A characterized biosphere flow associated with a given `Node` instance.
-
-    Parameters
-    ----------
-    flow_datapackage_id : int
-        The id that identifies the biosphere flow in the datapackage
-    flow_index : int
-        The matrix row index of the biosphere flow
-    activity_unique_id : int
-        The `Node.unique_id` of this instance of the emitting activity
-    activity_id : int
-        The id that identifies the emitting activity in the datapackage
-    activity_index : int
-        The matrix column index of the emitting activity
-    amount : float
-        The amount of the biosphere flow being emitting by this activity instance
-    score : float
-        The LCIA score for `amount` of this biosphere flow
-    """
-
-    flow_datapackage_id: int
-    flow_index: int
-    activity_unique_id: int
-    activity_id: int
-    activity_index: int
-    amount: float
-    score: float
-
-    def __lt__(self, other):
-        # Needed for sorting
-        return self.score < other.score
+from ..matrix_tools import guess_production_exchanges
+from .graph_objects import Node, Edge, Flow
+from .utils import CachingSolver, Counter
 
 
 class NewNodeEachVisitGraphTraversal:
@@ -726,217 +567,3 @@ class NewNodeEachVisitGraphTraversal:
             rows.append(x)
             vals.append(y)
         return rows, vals
-
-
-class AssumedDiagonalGraphTraversal(NewNodeEachVisitGraphTraversal):
-    @classmethod
-    def get_production_exchanges(cls, mapped_matrix: mu.MappedMatrix) -> (np.ndarray, np.ndarray):
-        """
-        Assume production exchanges are always on the diagonal instead of
-        examining matrix structure and input data.
-
-        Parameters
-        ----------
-        mapped_matrix : matrix_utils.MappedMatrix
-            A matrix and mapping data (from database ids to matrix indices)
-            from the ``matrix_utils`` library. Normally built automatically by
-            an ``LCA`` class. Should be the ``technosphere_matrix`` or
-            equivalent.
-
-        Returns
-        -------
-        (numpy.array, numpy.array)
-            The matrix row and column indices of the production exchanges.
-
-        """
-        length = mapped_matrix.matrix.shape[0]
-        return np.arange(length), np.arange(length)
-
-
-class StatefulGraphTraversal(NewNodeEachVisitGraphTraversal):
-    """
-    A stateful graph traversal that keeps track of which nodes have been
-    visited already, allowing for a partial and dynamic traversal of the graph.
-    """
-
-    def __init__(
-        self,
-        lca_object: LCA,
-        cutoff: Optional[float] = 5e-3,
-        biosphere_cutoff: Optional[float] = 1e-4,
-        max_calc: Optional[int] = 10000,
-        max_depth: Optional[int] = None,
-        skip_coproducts: Optional[bool] = False,
-        separate_biosphere_flows: Optional[bool] = True,
-        static_activity_indices: Optional[set[int]] = set(),
-        functional_unit_unique_id: Optional[int] = -1,
-    ):
-        """
-        See `NewNodeEachVisitGraphTraversal.calculate` for more information
-
-        Parameters
-        ----------
-        lca_object : bw2calc.LCA
-            Already instantiated `LCA` object with inventory and impact
-            assessment calculated.
-        cutoff : float
-            Cutoff value used to stop graph traversal. Fraction of total score,
-            should be in `(0, 1)`
-        biosphere_cutoff : float
-            Cutoff value used to determine if a separate biosphere node is
-            added. Fraction of total score.
-        max_calc : int
-            Maximum number of inventory calculations to perform
-        max_depth : int
-            Maximum depth in the supply chain for the initial traversal. Default is no maximum.
-        skip_coproducts : bool
-            Don't traverse co-production edges, i.e. production edges other
-            than the reference product
-        separate_biosphere_flows : bool
-            Add separate `Flow` nodes for important individual biosphere
-            emissions
-        static_activity_indices : set
-            A set of activity matrix indices which we don't want the graph to
-            traverse
-        functional_unit_unique_id : int
-            An integer id we can use for the functional unit virtual activity.
-            Shouldn't overlap any other activity ids. Don't change unless you
-            really know what you are doing.
-
-        """
-        super().__init__()
-        self.lca_object = lca_object
-        self.cutoff = cutoff
-        self.max_calc = max_calc
-        self.skip_coproducts = skip_coproducts
-        self.separate_biosphere_flows = separate_biosphere_flows
-        self.static_activity_indices = static_activity_indices
-        total_score = lca_object.score
-        if total_score == 0:
-            raise ValueError("Zero total LCA score makes traversal impossible")
-
-        self.cutoff_score = abs(total_score * cutoff)
-        self.biosphere_cutoff_score = abs(total_score * biosphere_cutoff)
-        self.technosphere_matrix = lca_object.technosphere_matrix
-        self.production_exchange_mapping = {
-            x: y for x, y in zip(*self.get_production_exchanges(lca_object.technosphere_mm))
-        }
-        self.heap, self.edges, self.flows = [], [], []
-        self.calculation_count = Counter()
-        self.characterized_biosphere = self.get_characterized_biosphere(lca_object)
-        self.caching_solver = CachingSolver(lca_object)
-
-        self.nodes: Dict[int, Node] = {
-            functional_unit_unique_id: Node(
-                unique_id=functional_unit_unique_id,
-                activity_datapackage_id=functional_unit_unique_id,
-                activity_index=functional_unit_unique_id,
-                reference_product_datapackage_id=functional_unit_unique_id,
-                reference_product_index=functional_unit_unique_id,
-                reference_product_production_amount=1.0,
-                depth=0,
-                supply_amount=1.0,
-                cumulative_score=lca_object.score,
-                direct_emissions_score=0.0,
-            )
-        }
-        self.edges: List[Edge] = []
-        self.flows = []
-        self.calculation_count = Counter()
-        self.visited_nodes = set()
-        self.metadata = dict()
-
-        self.traverse_edges(
-            consumer_index=functional_unit_unique_id,
-            consumer_unique_id=functional_unit_unique_id,
-            product_indices=[lca_object.dicts.product[key] for key in lca_object.demand],
-            product_amounts=lca_object.demand.values(),
-            lca=lca_object,
-            current_depth=0,
-            max_depth=max_depth,
-            calculation_count=self.calculation_count,
-            characterized_biosphere=self.characterized_biosphere,
-            matrix=self.technosphere_matrix,
-            edges=self.edges,
-            flows=self.flows,
-            nodes=self.nodes,
-            heap=self.heap,
-            caching_solver=self.caching_solver,
-            production_exchange_mapping=self.production_exchange_mapping,
-            separate_biosphere_flows=separate_biosphere_flows,
-            cutoff_score=self.cutoff_score,
-            static_activity_indices=self.static_activity_indices,
-            biosphere_cutoff_score=self.biosphere_cutoff_score,
-        )
-
-        self.traverse_from_node(functional_unit_unique_id, max_depth=max_depth)
-
-    def traversed_node(self, node_id: int) -> bool:
-        """
-        Check wither a node has been traversed
-
-        Parameters
-        ----------
-        node_id
-            node's unique id
-
-        Returns
-        -------
-        bool
-            indicates that the node has been traversed
-        """
-        return node_id in self.visited_nodes
-
-    def traverse_from_node(self, node_id: int, max_depth: Optional[int] = None) -> bool:
-        """
-        Traverse the graph starting from the specified node and exp
-        returning a boolean indicating if the node traversed already
-
-        Parameters
-        ----------
-        node_id
-            node's unique id
-        max_depth
-            max depth to traverse from this node, otherwise will default to `node.depth + 1`
-
-        Returns
-        -------
-        bool
-            indicates if the node was traversed
-        """
-        node = self.nodes[node_id]
-        if node.unique_id in self.visited_nodes:
-            return False
-
-        self.visited_nodes.add(node.unique_id)
-        self.traverse(
-            heap=[(0, node)],
-            nodes=self.nodes,
-            edges=self.edges,
-            flows=self.flows,
-            max_depth=node.depth + 1 if max_depth is None else max_depth,
-            max_calc=self.max_calc,
-            cutoff_score=self.cutoff_score,
-            characterized_biosphere=self.characterized_biosphere,
-            calculation_count=self.calculation_count,
-            static_activity_indices=self.static_activity_indices,
-            caching_solver=self.caching_solver,
-            production_exchange_mapping=self.production_exchange_mapping,
-            technosphere_matrix=self.technosphere_matrix,
-            lca=self.lca_object,
-            skip_coproducts=self.skip_coproducts,
-            separate_biosphere_flows=self.separate_biosphere_flows,
-            biosphere_cutoff_score=self.biosphere_cutoff_score,
-        )
-
-        self.flows.sort(reverse=True)
-
-        non_terminal_nodes = {edge.consumer_unique_id for edge in self.edges}
-        for node_id_key in self.nodes:
-            if node_id_key not in non_terminal_nodes:
-                terminal = True
-            else:
-                terminal = False
-            self.nodes[node_id_key].terminal = terminal
-
-        return True
